implementing conversion inside &. no more "&int32(x)".

+++ jl/linalg_blas.jl
@@ -8,7 +8,7 @@ for (fname, elty) in ((:dcopy_,:Float64), (:scopy_,:Float32),
             ccall(dlsym(_jl_libblas, $string(fname)),
                   Void,
                   (Ptr{Int32}, Ptr{$elty}, Ptr{Int32}, Ptr{$elty}, Ptr{Int32}),
-                  &int32(n), DX, &int32(incx), DY, &int32(incy))
+                  &n, DX, &incx, DY, &incy)
             return DY
         end
     end
@@ -41,7 +41,7 @@ for (fname, elty) in ((:ddot_,:Float64), (:sdot_,:Float32))
             ccall(dlsym(_jl_libblas, $string(fname)),
                   $elty,
                   (Ptr{Int32}, Ptr{$elty}, Ptr{Int32}, Ptr{$elty}, Ptr{Int32}),
-                  &int32(n), DX, &int32(incx), DY, &int32(incy))
+                  &n, DX, &incx, DY, &incy)
         end
     end
 end
@@ -65,7 +65,7 @@ for (fname, elty, ret_type) in ((:dnrm2_,:Float64,:Float64),
             ccall(dlsym(_jl_libblas, $string(fname)),
                   $ret_type,
                   (Ptr{Int32}, Ptr{$elty}, Ptr{Int32}),
-                  &int32(n), X, &int32(incx))
+                  &n, X, &incx)
         end
     end
 end
@@ -88,7 +88,7 @@ for (fname, elty) in ((:daxpy_,:Float64), (:saxpy_,:Float32),
             ccall(dlsym(_jl_libblas, $string(fname)),
                   Void,
                   (Ptr{Int32}, Ptr{$elty}, Ptr{$elty}, Ptr{Int32}, Ptr{$elty}, Ptr{Int32}),
-                  &int32(n), x, DA, &int32(incx), DY, &int32(incy))            
+                  &n, x, DA, &incx, DY, &incy)            
         end
     end
 end
@@ -114,10 +114,10 @@ for (fname, elty) in ((:dgemm_,:Float64), (:sgemm_,:Float32),
                   Ptr{$elty}, Ptr{$elty}, Ptr{Int32},
                   Ptr{$elty}, Ptr{Int32},
                   Ptr{$elty}, Ptr{$elty}, Ptr{Int32}),
-                 &uint8(transA), &uint8(transB), &int32(m), &int32(n), &int32(k),
-                 &alpha, A, &int32(lda),
-                 B, &int32(ldb),
-                 &beta, C, &int32(ldc))
+                 &transA, &transB, &m, &n, &k,
+                 &alpha, A, &lda,
+                 B, &ldb,
+                 &beta, C, &ldc)
        end
 
    end
@@ -222,10 +222,10 @@ for (fname, elty) in ((:dgemv_,:Float64), (:sgemv_,:Float32),
                   Ptr{$elty}, Ptr{$elty}, Ptr{Int32},
                   Ptr{$elty}, Ptr{Int32},
                   Ptr{$elty}, Ptr{$elty}, Ptr{Int32}),
-                 trans, &int32(m), &int32(n),
-                 &alpha, A, &int32(lda),
-                 X, &int32(incx),
-                 &beta, Y, &int32(incy))
+                 trans, &m, &n,
+                 &alpha, A, &lda,
+                 X, &incx,
+                 &beta, Y, &incy)
        end
 
    end
